requirements.txt
# Core Framework
Flask==3.0.0
flask-cors==4.0.0

# Database
pymongo==4.6.0

# AI/NLP
sentence-transformers>=2.2.0
torch>=2.0.0
numpy>=1.24.0

# Configuration
python-dotenv==1.0.0